- S/L fixed

+++ YRpp/FacingClass.h
@@ -104,7 +104,7 @@ public:
 		if (!Is_Rotating())
 			return false;
 
-		return Difference().Raw < 0;
+		return static_cast<short>(Difference().Raw) < 0;
 	}
 
 	bool Is_Rotating_G() const
@@ -112,7 +112,7 @@ public:
 		if (!Is_Rotating())
 			return false;
 
-		return Difference().Raw > 0;
+		return static_cast<short>(Difference().Raw) > 0;
 	}
 
 	DirStruct Difference() const
